Move `usart::Clock` impl blocks to `usart`
This reduces code duplication between the different targets, and makes
them a bit easier to find.
